Storing in EEPROM Clearing EEPROM

+++ include/Config.h
@@ -3,17 +3,17 @@ public:
   struct Options {
     bool buzzerOn = true;
     bool lcdAlwaysOn = true;
-    float direction = -1;
+    double direction = -1;
   };
   enum class MenuOptions : short { notImplemented = 0 };
   Config();
   void show();
   void setBuzzerOn(bool isBuzzerOn);
   void setLcdAlwaysOn(bool isLcdAlwaysOn);
-  void setDirection(float direction);
+  void setDirection(double direction);
   bool isBuzzerOn();
   bool isLcdAlwaysOn();
-  float getDirection();
+  double getDirection();
 
 private:
   MenuOptions _menuOption = MenuOptions::notImplemented;
